backend/pyproject.toml
[project]
name = "nomad-backend"
version = "0.1.0"
description = "Nomad Notes Django backend service"
authors = [{ name = "Nomad Team" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "django>=5.1,<6",
  "djangorestframework>=3.15,<4",
  "djangorestframework-simplejwt>=5.3,<6",
  "django-environ>=0.11,<1",
  "psycopg[binary]>=3.2,<4",
  "drf-spectacular>=0.27,<0.28"
]

[project.optional-dependencies]
dev = [
  "pytest-django>=4.8,<5",
  "pytest>=8.3,<9",
  "ruff>=0.7,<0.8"
]

[tool.uv]
package = true

[tool.uv.pip]
index-url = "https://pypi.org/simple"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "nomad_backend.settings"
python_files = ["tests.py", "test_*.py"]
addopts = "-s"
